+++ packages/familie-backend/lib/typer.ts
@@ -1,5 +1,6 @@
 import { Express, Request } from 'express';
 import { Registry } from 'prom-client';
+import { IOIDCStrategyOptionWithRequest } from 'passport-azure-ad';
 
 export interface IFamilieBackend {
     app: Express;
@@ -24,6 +25,8 @@ export interface ITokenRequest {
     clientId: string;
     clientSecret: string;
     redirectUrl: string;
-    tokenUri: string;
     scope: string;
+    tokenUri: string;
 }
+
+export { IOIDCStrategyOptionWithRequest };