+++ packages/familie-backend/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.9...@navikt/familie-backend@1.1.0) (2019-11-19)
+
+
+### Features
+
+* **backend:** legger til router med rutene til azure flyt og brukerprofil. Rydder opp og setter opp ([5f9b6ca](https://github.com/navikt/familie-felles-frontend/commit/5f9b6cab34bd09f1b72c70e87ba0e75fa33a0eee))
+
+
+
+
+
+## [1.0.9](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.8...@navikt/familie-backend@1.0.9) (2019-11-18)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+
+
+
+
+## [1.0.8](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.7...@navikt/familie-backend@1.0.8) (2019-11-18)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+
+
+
+
+## [1.0.7](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.6...@navikt/familie-backend@1.0.7) (2019-11-18)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+## [1.0.6](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.5...@navikt/familie-backend@1.0.6) (2019-11-13)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+## [1.0.5](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.4...@navikt/familie-backend@1.0.5) (2019-11-13)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
+## [1.0.4](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.3...@navikt/familie-backend@1.0.4) (2019-11-13)
+
+**Note:** Version bump only for package @navikt/familie-backend
+
 ## [1.0.3](https://github.com/navikt/familie-felles-frontend/compare/@navikt/familie-backend@1.0.1...@navikt/familie-backend@1.0.3) (2019-11-13)
 
 **Note:** Version bump only for package @navikt/familie-backend